Reorganize user_lang modules
