+++ axum-extra/src/user_lang/mod.rs
@@ -1,10 +1,11 @@
+//! TBD
+
 mod config;
 mod lang;
 mod source;
-
-/// TBD
-pub mod sources;
+mod sources;
 
 pub use config::*;
 pub use lang::*;
 pub use source::*;
+pub use sources::*;
