+++ axum-extra/src/user_lang/sources/header.rs
@@ -1,7 +1,6 @@
-use std::cmp::Ordering;
-
-use crate::extract::UserLanguageSource;
+use crate::user_lang::UserLanguageSource;
 use axum::async_trait;
+use std::cmp::Ordering;
 
 /// TBD
 #[derive(Debug, Clone)]
