+++ axum-extra/src/user_lang/sources/path.rs
@@ -1,15 +1,15 @@
-use crate::extract::UserLanguageSource;
-use axum::{async_trait, extract::Query, RequestPartsExt};
+use crate::user_lang::UserLanguageSource;
+use axum::{async_trait, extract::Path, RequestPartsExt};
 use std::collections::HashMap;
 
 /// TBD
 #[derive(Debug, Clone)]
-pub struct QuerySource {
+pub struct PathSource {
     /// TBD
     name: String,
 }
 
-impl QuerySource {
+impl PathSource {
     /// TBD
     pub fn new(name: impl Into<String>) -> Self {
         Self { name: name.into() }
@@ -17,13 +17,13 @@ impl QuerySource {
 }
 
 #[async_trait]
-impl UserLanguageSource for QuerySource {
+impl UserLanguageSource for PathSource {
     async fn languages_from_parts(&self, parts: &mut http::request::Parts) -> Vec<String> {
-        let Ok(query) = parts.extract::<Query<HashMap<String, String>>>().await else {
+        let Ok(path) = parts.extract::<Path<HashMap<String, String>>>().await else {
             return vec![];
         };
 
-        let Some(lang) = query.get(self.name.as_str()) else {
+        let Some(lang) = path.get(self.name.as_str()) else {
             return vec![];
         };
