+++ axum-extra/src/user_lang/sources/query.rs
@@ -1,15 +1,15 @@
-use crate::extract::UserLanguageSource;
-use axum::{async_trait, extract::Path, RequestPartsExt};
+use crate::user_lang::UserLanguageSource;
+use axum::{async_trait, extract::Query, RequestPartsExt};
 use std::collections::HashMap;
 
 /// TBD
 #[derive(Debug, Clone)]
-pub struct PathSource {
+pub struct QuerySource {
     /// TBD
     name: String,
 }
 
-impl PathSource {
+impl QuerySource {
     /// TBD
     pub fn new(name: impl Into<String>) -> Self {
         Self { name: name.into() }
@@ -17,13 +17,13 @@ impl PathSource {
 }
 
 #[async_trait]
-impl UserLanguageSource for PathSource {
+impl UserLanguageSource for QuerySource {
     async fn languages_from_parts(&self, parts: &mut http::request::Parts) -> Vec<String> {
-        let Ok(path) = parts.extract::<Path<HashMap<String, String>>>().await else {
+        let Ok(query) = parts.extract::<Query<HashMap<String, String>>>().await else {
             return vec![];
         };
 
-        let Some(lang) = path.get(self.name.as_str()) else {
+        let Some(lang) = query.get(self.name.as_str()) else {
             return vec![];
         };
